+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # Change Log
 
-- 2019-11-05 (BF) initial version
+- 2019-12-10 (BF) split Data.Pad in Data.Pad, Data.Pass, Data.PadActions
 - 2019-11-18 (BF) add transmit :: MonadIO m => PAD -> ReportT m String
+- 2019-11-05 (BF) initial version